implement randBetween()

+++ src/com/codecool/crs/Vehicle.java
@@ -6,7 +6,7 @@ class Vehicle {
     String name;
 
     Integer getDistanceTraveled(){
-        return (Integer) distanceTraveled;
+        return distanceTraveled;
     }
 
     void moveForAnHour(){}
